Disregard dtx when pruning IC history
This patch aims to fix the memory leak issue #10314 .

Previously, the IC history table got pruned only after begining a
new distributed transaction (dtx). As a result, when a transaction
contains many commands, or when a session never begins any
distributed transaction, the history will never get pruned. This can
cause memory leak.

This patch fixes this issue by reverting this part of code to the
last version before dependency on distributed transaction was
introduced. After that, we will only a constant-length IC history.
This ensures that

- The history table takes only constant amount of memory, and
- The history is long enough to handle mismatched packets.