Support building standalone management utils
Previously, it was required to set PYTHONPATH before using the GP
management utils. Since PYTHONPATH will affect where to lookup
Python packages, having this dependency make it inconvenient if
the user wants to do other things, such as data science, in Python
on the same machine.

This patch adds support for building the GP management utils into
standalone executables with cx_freeze so that dependency on
PYTHONPATH can be removed.

Note that this patch does not affect what and how we release. The
feature is completely optional and can be configured with
--enable-standalone-utils. This patch is expected to not change
any bahavior of any util.

Known limitations and workarounds:
- cx_freeze does not create links for the included files. Thus, we
  need to copy them to $GPHOME/bin or $GPHOME/sbin in Makefile
- after the utils are built into standalone binaries, `__file__`
  will no longer be valid. Thus, we need to replace them with
  appropriate script names or paths.
- gppylib uses `python -c` to call Python functions directly during
  remote execution. Since PYTHONPATH is not set, we need to add the
  path manually before calling the functions.