Support resetting GUC when cluster down
Even though gpconfig tries its best to validate GUC setting, there
are still some cases a GUC will be set to an invalid value, e.g. because
of a typo or a bug. For safety, an invalid GUC setting might
prevent the cluster from starting. Since currently gpconfig
does not support resetting GUCs when cluster is down, user will need
to manually reset the invalid GUC in this case, which can be
cumbersome.

To overcome this limitation, this patch introduces a new util to reset a
GUC when cluster is down. This is achieved by reading the log of
previous log of goconfig when running it with --verbose.

The new util is designed for developers only and will not be installed to
`$GPHOME`. Therefore, this patch will not have any impact on what
we release.