Replace SCATTER BY 1 with column name
In Table-Value Expressions, `SCATTER BY 1` will redistribute the data
by constant number `1` instead of the first column. However, the
example in the doc implies that `SCATTER BY 1` will set
the distribution key to the **first** column, which is not true.

This patch fixes this by changing the `SCATTER BY 1` to `SCATTER BY
<column_name>` in the example of the doc to show how to redistribute
data by a column.